pyproject.toml
[tool.poetry]
name = "MLutilities"
version = "1.0.0-dev"
description = "Library to teach Machine Learning"
authors = ["omarcastano <omar.castano@udea.edu.co>", "deoache <daniel.ocampoh@udea.edu.co>"]
packages = [
    {include = "MLutilities"}
]

[tool.poetry.dependencies]
python = ">=3.7.13, <=3.9.0"
statsmodels = "^0.13.2"
plotly = "^5.10.0"
scikit-plot = "^0.3.7"
yellowbrick = "^1.5"
mlxtend = ">=0.19.0"
numpy = "^1.21.6"
lofo-importance = "^0.3.2"

[tool.poetry.dev-dependencies]
numpy = "^1.21.6"
matplotlib = "^3.2.2"
ipywidgets = "^7.7.0"
scipy = "^1.4.1"
jupyterlab = "^3.4.3"
jupyterlab-vim = "^0.15.1"
pandas = "^1.3.5"
seaborn = "^0.11.2"
ipykernel = "^6.15.1"
plotly = "^5.9.0"
black = "^22.6.0"
sklearn = "^0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
